backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
numpy==1.26.3
pydantic==2.5.3
python-multipart==0.0.6
aiohttp==3.9.1
httpx==0.27.0
beautifulsoup4==4.12.3
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
lxml==5.1.0
html5lib==1.1
yfinance==0.2.36
